OpXxx is part of public API
