+++ src/lib.rs
@@ -169,6 +169,20 @@ pub use error::Error;
 pub use error::ErrorKind;
 pub use error::Result;
 
+mod ops;
+pub use ops::OpAbortMultipart;
+pub use ops::OpCompleteMultipart;
+pub use ops::OpCreate;
+pub use ops::OpCreateMultipart;
+pub use ops::OpDelete;
+pub use ops::OpList;
+pub use ops::OpPresign;
+pub use ops::OpRead;
+pub use ops::OpStat;
+pub use ops::OpWrite;
+pub use ops::OpWriteMultipart;
+pub use ops::PresignOperation;
+
 // Public modules, they will be accessed via `opendal::layers::Xxxx`
 pub mod layers;
 pub use layers::Layer;
