+++ src/ops/op_read.rs
@@ -12,6 +12,28 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-/// Reply fro `delete` operation
+use crate::raw::*;
+
+/// Args for `read` operation.
 #[derive(Debug, Clone, Default)]
-pub struct RpDelete {}
+pub struct OpRead {
+    br: BytesRange,
+}
+
+impl OpRead {
+    /// Create a default `OpRead` which will read whole content of object.
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// Create a new OpRead with range.
+    pub fn with_range(mut self, range: BytesRange) -> Self {
+        self.br = range;
+        self
+    }
+
+    /// Get range from OpRead.
+    pub fn range(&self) -> BytesRange {
+        self.br
+    }
+}
