+++ src/raw/ops/mod.rs
@@ -25,33 +25,21 @@ mod operation;
 pub use operation::Operation;
 
 mod op_create;
-pub use op_create::OpCreate;
 pub use op_create::RpCreate;
 mod op_delete;
-pub use op_delete::OpDelete;
 pub use op_delete::RpDelete;
 mod op_list;
-pub use op_list::OpList;
 pub use op_list::RpList;
 mod op_presign;
-pub use op_presign::OpPresign;
-pub use op_presign::PresignOperation;
 pub use op_presign::PresignedRequest;
 pub use op_presign::RpPresign;
 mod op_read;
-pub use op_read::OpRead;
 pub use op_read::RpRead;
 mod op_stat;
-pub use op_stat::OpStat;
 pub use op_stat::RpStat;
 mod op_write;
-pub use op_write::OpWrite;
 pub use op_write::RpWrite;
 mod op_multipart;
-pub use op_multipart::OpAbortMultipart;
-pub use op_multipart::OpCompleteMultipart;
-pub use op_multipart::OpCreateMultipart;
-pub use op_multipart::OpWriteMultipart;
 pub use op_multipart::RpAbortMultipart;
 pub use op_multipart::RpCompleteMultipart;
 pub use op_multipart::RpCreateMultipart;
