+++ src/raw/ops/op_list.rs
@@ -12,17 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-/// Args for `list` operation.
-#[derive(Debug, Clone, Default)]
-pub struct OpList {}
-
-impl OpList {
-    /// Create a new `OpList`.
-    pub fn new() -> Self {
-        Self {}
-    }
-}
-
 /// Reply for `list` operation.
 #[derive(Debug, Clone, Default)]
 pub struct RpList {}
