+++ src/raw/ops/op_multipart.rs
@@ -14,17 +14,6 @@
 
 use crate::ObjectPart;
 
-/// Args for `create_multipart` operation.
-#[derive(Debug, Clone, Default)]
-pub struct OpCreateMultipart {}
-
-impl OpCreateMultipart {
-    /// Create a new `OpCreateMultipart`.
-    pub fn new() -> Self {
-        Self {}
-    }
-}
-
 /// Reply for `create_multipart` operation.
 #[derive(Debug, Clone, Default)]
 pub struct RpCreateMultipart {
@@ -42,40 +31,6 @@ impl RpCreateMultipart {
     /// Get the upload_id.
     pub fn upload_id(&self) -> &str {
         &self.upload_id
-    }
-}
-
-/// Args for `write_multipart` operation.
-#[derive(Debug, Clone, Default)]
-pub struct OpWriteMultipart {
-    upload_id: String,
-    part_number: usize,
-    size: u64,
-}
-
-impl OpWriteMultipart {
-    /// Create a new `OpWriteMultipart`.
-    pub fn new(upload_id: String, part_number: usize, size: u64) -> Self {
-        Self {
-            upload_id,
-            part_number,
-            size,
-        }
-    }
-
-    /// Get upload_id from option.
-    pub fn upload_id(&self) -> &str {
-        &self.upload_id
-    }
-
-    /// Get part_number from option.
-    pub fn part_number(&self) -> usize {
-        self.part_number
-    }
-
-    /// Get size from option.
-    pub fn size(&self) -> u64 {
-        self.size
     }
 }
 
@@ -111,55 +66,9 @@ impl RpWriteMultipart {
     }
 }
 
-/// Args for `complete_multipart` operation.
-#[derive(Debug, Clone, Default)]
-pub struct OpCompleteMultipart {
-    upload_id: String,
-    parts: Vec<ObjectPart>,
-}
-
-impl OpCompleteMultipart {
-    /// Create a new `OpCompleteMultipart`.
-    pub fn new(upload_id: String, parts: Vec<ObjectPart>) -> Self {
-        Self { upload_id, parts }
-    }
-
-    /// Get upload_id from option.
-    pub fn upload_id(&self) -> &str {
-        &self.upload_id
-    }
-
-    /// Get parts from option.
-    pub fn parts(&self) -> &[ObjectPart] {
-        &self.parts
-    }
-}
-
 /// Reply for `complete_multipart` operation.
 #[derive(Debug, Clone, Default)]
 pub struct RpCompleteMultipart {}
-
-/// Args for `abort_multipart` operation.
-///
-/// The path must be normalized.
-#[derive(Debug, Clone, Default)]
-pub struct OpAbortMultipart {
-    upload_id: String,
-}
-
-impl OpAbortMultipart {
-    /// Create a new `OpAbortMultipart`.
-    ///
-    /// If input path is not a file path, an error will be returned.
-    pub fn new(upload_id: String) -> Self {
-        Self { upload_id }
-    }
-
-    /// Get upload_id from option.
-    pub fn upload_id(&self) -> &str {
-        &self.upload_id
-    }
-}
 
 /// Reply for `abort_multipart` operation.
 #[derive(Debug, Clone, Default)]
