+++ src/raw/ops/op_write.rs
@@ -12,42 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-/// Args for `write` operation.
-#[derive(Debug, Clone, Default)]
-pub struct OpWrite {
-    size: u64,
-    content_type: Option<String>,
-}
-
-impl OpWrite {
-    /// Create a new `OpWrite`.
-    ///
-    /// If input path is not a file path, an error will be returned.
-    pub fn new(size: u64) -> Self {
-        Self {
-            size,
-            content_type: None,
-        }
-    }
-
-    /// Set the content type of option
-    pub fn with_content_type(self, content_type: &str) -> Self {
-        Self {
-            size: self.size(),
-            content_type: Some(content_type.to_string()),
-        }
-    }
-
-    /// Get size from option.
-    pub fn size(&self) -> u64 {
-        self.size
-    }
-    /// Get the content type from option
-    pub fn content_type(&self) -> Option<&str> {
-        self.content_type.as_deref()
-    }
-}
-
 /// Reply for `write` operation.
 #[derive(Debug, Clone, Default)]
 pub struct RpWrite {
